tracing/requirements.txt
Flask==2.1.2
Werkzeug==2.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-instrumentation-grpc==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-urllib==0.43b0
opentelemetry-semantic-conventions==0.43b0
requests
loguru
